[Update] notice popup function
