+++ app/functuion.py
@@ -24,10 +24,22 @@ def session_check():
         return redirect(url_for('login'))
 
 def get_post(id):
+    # post 정보
     col_post = db.get_collection('post')
     post = col_post.find_one({'_id':ObjectId(id)})
     post['_id'] = str(post['_id'])
+    # 댓글 정보
+    col_comment = db.get_collection('comment')
+    comment_list = list(col_comment.find({'post_id': id}))
+    for i in range(len(comment_list)):
+        comment_list[i]['_id'] = str(comment_list[i]['_id']) 
+
+    post['comment_info'] = comment_list
     return post
+
+def get_comment(id):
+    
+    return comment_list
 
 def s3_connection():
     try:
